Edit: AddRecordView 주소 받아오기 에러 수정, 파일 분리 - 13.4
